Added repo and service for modelClasses

+++ Backend/jackpot/src/main/java/com/fidelity/jackpot/model/TradeHistory.java
@@ -15,7 +15,7 @@ import java.sql.Timestamp;
 public class TradeHistory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer TradeHistoryID;
+    private Long TradeHistoryID;
 
     private String fundName;
