+++ Backend/jackpot/src/main/java/com/fidelity/jackpot/model/User.java
@@ -16,7 +16,7 @@ import java.util.Set;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer userID;
+    private Long userID;
 
     private String firstName;
